Remove an Android change that is causing a test to fail
We have a patch in KeyAgreementSpi that changes the set of exceptions that
are caught from within Bouncy Castle from Exception to only
IllegalStateException.

This patch was introduced in the upgrade to BC 1.56, and there's no
explanation why it was introduced.  Prior to that upgrade, upstream caught
no exceptions and we caught IllegalStateException.  (See aosp/318406.)

This shouldn't cause any problems for users, as the set of exceptions that
is thrown by the internal key agreement class is an implementation detail
and InvalidKeyException and IllegalStateException are the only documented
exceptions thrown from KeyAgreement#doPhase() (which is the public API this
backs), so any other exception type would be unexpected anyway.

Bug: 36712087
Test: cts -m CtsLibcoreTestCases
Change-Id: Idb18fe0cf7bcf5a86e8805c362941528249aad2a