Update change markers in external/bouncycastle.
For the most part, this is just changing markers that had no explanatory
text and adding an explanation.  Some more detail on two of the common
explanations:

"Unsupported algorithms": We choose not to ship a whole lot of the
various crypto algorithms that Bouncy Castle supports.  Justfying each
individual choice to disable a given algorithm is out of place in the
code and also doesn't have a natural home (since each algorithm is
disabled in upwards of a dozen places), so just mark where algorithms
are disabled due to being unsupported.

"Unknown reason": I looked in the history and couldn't find any
justification for the change.  Generally these changes were made in
commits with descriptions like "bouncycastle 1.49 upgrade" from multiple
years ago.  I don't want to change behavior in this change, so I left
these changes as they were; these can be revisited in the future if
desired.

This change also reverts some diffs from upstream, both inside disabled
blocks of code (which is irrelevant) and ones that are active but have
no behavioral change (for example, an added package-visible constructor
that isn't called by any class in the package).  There're also a number
of whitespace-only changes that revert diffs from upstream.

Bug: 63092064
Test: cts -m CtsLibcoreTestCases
Change-Id: I27c61063eda849455bc93edaf180acefb5eed217